[feat] manage card flip
problem: you need a sense of order, you need to lock a card as flipped only if the matching card immediately follows
solution: when flipping a card remove it to avoid multiple clicks and show one in its place, animated. At the same time overlay a card _on the matching_ counterpart. Listen to a click on this specific overlay to lock the original in place; selecting any other card — see resets — flips the original card back, reintroduces the original card and removes the overlay