[feat] add target overlays
the problem: detecting when the projectile fired by the spaceship point of view should hit a target above
the (rough) solution: add space around the target to make up for the gap in vertical dimension; compute the time in which 1. the target is in the hit area and 2. the time it takes for the target to return to the hit area; with these values add and remove the overlay in conjunction with the hit area; detecting a click on this overlay should result in the target being shot; detecting a click on the rectangle beneat should result in a missed shot