refactor puzzle component
instead of styles.js have shapes.js export an object where each key has two properties: with styles repeat the transformations, with padding set the whitespace to display the shapes (increments of 0.5) || in Tangram.svelte pick a shape at random based on a key || in Tangram.svelte make the key **and** the padding as two props, but extract the padding from the object if not specified || the key & padding combination means that 1. static illustrations have the necessary whitespace and 2. when you change the key the padding remains the same. If you want  to change the padding do so from the parent scope